Refactor: abstract `HistoryEntry` into own component
